# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.BlockchainResponse:
    properties:
      errors:
        example: No Error Message
        type: string
      returned: {}
      success:
        example: true
        type: boolean
    type: object
...
  response.Meta:
    properties:
      timestamp:
        type: string
      version:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      error: {}
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: DID Backend API
  title: DID API
  version: "1.0"
paths:
...
  /did-root/did-index-map:
    get:
      consumes:
      - application/json
      description: Maps a DID index to its corresponding DID hash
      parameters:
      - description: DID index (hex-encoded big integer with 0x prefix)
        example: "0x1"
        in: query
        name: did_index
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: DID hash retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.BlockchainResponse'
              type: object
        "400":
          description: Invalid request - missing or malformed did_index parameter
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error - mapping retrieval failed or index not
            found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get DID Hash by Index
      tags:
      - DID Root GET
  /did-root/did-index-map-reverse:
    get:
      consumes:
      - application/json
      description: Reverse maps a DID hash to its corresponding DID index
      parameters:
      - description: DID hash (32-byte hex string with 0x prefix)
        example: 0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef
        in: query
        name: did_hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: DID index retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.BlockchainResponse'
              type: object
        "400":
          description: Invalid request - missing or malformed did_hash parameter
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error - reverse mapping retrieval failed or
            hash not found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get DID Index by Hash
      tags:
      - DID Root GET
  /did-root/get-claim:
    get:
      consumes:
      - application/json
      description: Retrieves a specific verifiable claim associated with a DID
      parameters:
      - description: DID hash (32-byte hex string with 0x prefix)
        example: 0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef
        in: query
        name: did_hash
        required: true
        type: string
      - description: Claim identifier (32-byte hex string with 0x prefix)
        example: 0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890
        in: query
        name: claim_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Claim retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.BlockchainResponse'
              type: object
        "400":
          description: Invalid request - missing parameters or invalid hash format
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error - claim retrieval failed or not found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Claim by DID and Claim ID
      tags:
      - DID Root GET
  /did-root/get-claims-by-did:
    get:
      consumes:
      - application/json
      description: Retrieves all verifiable claims associated with a specific DID
      parameters:
      - description: DID hash (32-byte hex string with 0x prefix)
        example: 0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef
        in: query
        name: did_hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Claims retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.BlockchainResponse'
              type: object
        "400":
          description: Invalid request - missing or malformed did_hash parameter
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error - claims retrieval failed
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get All Claims for a DID
      tags:
      - DID Root GET
  /did-root/get-data:
    get:
      consumes:
      - application/json
      description: Retrieves data associated with a specific hash from the DID Root
        contract
      parameters:
      - description: Data hash (32-byte hex string with 0x prefix)
        example: 0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef
        in: query
        name: did_hash
        required: true
        type: string
      - description: key string
        example: email
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.BlockchainResponse'
              type: object
        "400":
          description: Invalid request - missing or malformed hash parameter
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error - contract call failed or network issues
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Data by Hash
      tags:
      - DID Root GET
  /did-root/get-did-key-data-by-index:
    get:
      consumes:
      - application/json
      description: Retrieves a specific key associated with a DID using its index
        position
      parameters:
      - description: DID hash (32-byte hex string with 0x prefix)
        example: 0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef
        in: query
        name: did_hash
        required: true
        type: string
      - description: Key index (unsigned integer as string)
        example: "0"
        in: query
        name: index
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Key retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.BlockchainResponse'
              type: object
        "400":
          description: Invalid request - missing parameters, invalid hash format,
            or invalid index
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error - key retrieval failed or index out of
            bounds
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get DID Key by Index
      tags:
      - DID Root GET
  /did-root/get-did-key-data-count:
    get:
      consumes:
      - application/json
      description: Returns the total number of keys associated with a specific DID
      parameters:
      - description: DID hash (32-byte hex string with 0x prefix)
        example: 0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef
        in: query
        name: did_hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Key count retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.BlockchainResponse'
              type: object
        "400":
          description: Invalid request - missing or malformed did_hash parameter
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error - count retrieval failed
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get DID Key Data Count
      tags:
      - DID Root GET
  /did-root/get-key:
    get:
      consumes:
      - application/json
      description: Retrieves a specific cryptographic key associated with a DID using
        the key's hashed data
      parameters:
      - description: DID hash (32-byte hex string with 0x prefix)
        example: 0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef
        in: query
        name: did_hash
        required: true
        type: string
      - description: Hashed key data (32-byte hex string with 0x prefix)
        example: 0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890
        in: query
        name: key_data_hashed
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Key retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.BlockchainResponse'
              type: object
        "400":
          description: Invalid request - missing parameters or invalid hash format
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error - key retrieval failed or not found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Key by DID and Key Hash
      tags:
      - DID Root GET
  /did-root/get-keys-by-did:
    get:
      consumes:
      - application/json
      description: Retrieves all cryptographic keys associated with a specific DID
      parameters:
      - description: DID hash (32-byte hex string with 0x prefix)
        example: 0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef
        in: query
        name: did_hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Keys retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.BlockchainResponse'
              type: object
        "400":
          description: Invalid request - missing or malformed did_hash parameter
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error - keys retrieval failed
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get All Keys for a DID
      tags:
      - DID Root GET
  /did-root/get-original-key:
    get:
      consumes:
      - application/json
      description: Retrieves the original key data using a key code identifier
      parameters:
      - description: Key code identifier (32-byte hex string with 0x prefix)
        example: 0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef
        in: query
        name: key_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Original key retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.BlockchainResponse'
              type: object
        "400":
          description: Invalid request - missing or malformed key_code parameter
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error - key retrieval failed or not found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Original Key by Key Code
      tags:
      - DID Root GET
...
swagger: "2.0"